requirements.txt
# Google Cloud dependencies
google-cloud-storage==2.14.0
google-cloud-bigquery==3.13.0
google-cloud-secret-manager==2.18.0
google-cloud-logging==3.8.0

# Data processing
pandas==2.1.4
pyarrow==14.0.2

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

